sentiment_research/requirements.txt
# Aspect-Based Sentiment Analysis Research
# Python 3.9+

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0

# NLP
nltk>=3.8.0

# Optional: For Jupyter notebooks
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
